pyproject.toml
[project]
name = "sandol-auth-relay"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "asyncio>=4.0.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "pyjwt>=2.10.1",
    "python-keycloak>=5.8.1",
    "uvicorn>=0.35.0",
]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "docs",
    "alembic"
]

line-length = 88
indent-width = 4

target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "C", "D", "R", "B", "N", "S", "PL"]
ignore = ["E203", "E501", "D415", "D403"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_|__)$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[dependency-groups]
dev = [
    "ruff>=0.12.9",
]

[tool.ruff.lint.pydocstyle]
convention = "google"
